+++ website/src/utils/seo.ts
@@ -5,7 +5,6 @@ const baseKeywords = [
   '微信小程序',
   'uni-app',
   'taro',
-  'rax',
   'mpx',
 ]